requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
openai
faiss-cpu
pypdf
tiktoken